Card data Custom Inspector, Install sprite to card template...
1. Install all card template with imported sprite, and adjust the code logic of each template data loader script.
( Using  inherit ). Also, adjust all card template size, position and some setting.
2. Now when the card is weapon or armor, it will show it's AP, CR, DP,DSP value with some card template, but some value don't effect to player.
3. Write the Card data scriptable Object custom inspector, now you can create card data more efficiently.
4. Add the spells effect value storage variable to spells card data, but now it can only have one variable.
5. Add some bonus value to equipment card data, but it didn't fully connect to status system yet.
6. Adjust the equipment slot UI display due to the add of bonus variable.
7. Adjust size, position of some UI.

Add skybox.
Fix the draw card bug.

+++ Assets/Scripts/ScriptableObject/Status/PlayerStatus.cs
@@ -25,9 +25,9 @@ public class PlayerStatus : ScriptableObject
     public int baseMaxMana = 10; // will not increase with lv.
     //[HideInInspector] public int currentMaxMana; 
     //[HideInInspector] public int currentMana;
-    public int baseManaRT = 10; 
+    public float baseManaRT = 10; 
 
-    public int baseMaxSpeed = 1; // will not increase with lv.
+    public int baseMaxSpeed = 1; 
     //[HideInInspector] public int currentMaxSpeed;
 
     public int Money = 0;
